+++ tests/dataset/util/simple_dataset.rs
@@ -2,15 +2,15 @@ use nove::dataset::Dataset;
 
 /// A simple dataset for testing.
 ///
-/// The type of item in the dataset is `u64`. It contains
+/// The type of item in the dataset is `usize`. It contains
 /// 100 items, each item is the index of the item.
 pub struct SimpleDataset {}
 impl Dataset for SimpleDataset {
-    type Item = u64;
-    fn get(&self, index: u64) -> Self::Item {
+    type Item = usize;
+    fn get(&self, index: usize) -> Self::Item {
         index
     }
-    fn len(&self) -> u64 {
+    fn len(&self) -> usize {
         100
     }
 }
